# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TTR ONE — ERP Platform API
  version: 0.1.0
  description: |
    REST API for **TTR ONE**, a multi-tenant SaaS ERP for Uzbekistan
    (Node/Fastify + PostgreSQL/Prisma backend, Nuxt 3 SPA frontend).

    ## Conventions
    - **Base URL**: all endpoints live under `/api/v1`.
    - **Auth**: JWT Bearer. Obtain an `accessToken` from `POST /auth/login`
      (15-minute TTL) and rotate it via `POST /auth/refresh`.
    - **Money**: all monetary amounts are integers in **minor units (tiyin)**,
      e.g. `1 000 000` = 10 000.00 UZS. Fields end in `Minor`.
    - **Quantities**: decimal strings, bounded to 6 decimal places.
    - **Multi-tenancy**: every record is scoped to the caller's tenant
      (derived from the token); there is no tenant id in the path.
    - **Errors**: uniform envelope `{ "error": { "code": "...", "message": "..." } }`.
    - **Pagination**: list endpoints accept `page`, `pageSize` (max 200),
      `sort`, `order` (`asc`/`desc`), `search`, and return a `meta` object
      `{ page, pageSize, total, totalPages }`.
  contact:
    name: TTR ONE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - apiKeyAuth: []
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Self-registration — create a tenant + owner and log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Tenant created and session issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email + password (optional tenant slug and MFA code)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate a refresh token for a new access + refresh pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/me:
    get:
      tags: [auth]
      summary: Current authenticated user + permissions
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change password (invalidates all sessions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password-reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                tenant: { type: string }
      responses:
        '200':
          description: Always ok (does not reveal whether the email exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
                  devResetToken: { type: string, description: Dev only }
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Set a new password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /auth/mfa/enable:
    post:
      tags: [auth]
      summary: Confirm and enable MFA with a TOTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties: { code: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /auth/pin/set:
    post:
      tags: [auth]
      summary: Set a 4–8 digit mobile PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties: { pin: { type: string, pattern: '^\d{4,8}$' } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token from POST /auth/login, sent as `Authorization: Bearer <token>`.'
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Developer API key (ttr_live_…) created in the app under Управление → API-ключи. Recommended for server-to-server integrations. May also be sent as `Authorization: Bearer ttr_live_…`.'
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean, example: true } }
    ValidationError:
      description: Validation or business-rule error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials/token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, nullable: true }
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        tenant: { type: string, description: 'Slug, when the email exists in multiple tenants' }
        mfaCode: { type: string }
    RegisterRequest:
      type: object
      required: [companyName, fullName, email, password]
      properties:
        companyName: { type: string }
        industry: { type: string, description: 'Niche key, e.g. manufacturing, retail' }
        fullName: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        plan: { type: string }
        modules:
          type: array
          items: { type: string }
        bill: { type: object, description: 'Optional legal/billing requisites' }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tenant:
          type: object
          properties:
            id: { type: string }
            slug: { type: string }
            name: { type: string }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        tenantId: { type: string }
        roles:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
        platformAdmin: { type: boolean }
...